Refactor caption display and transcription handling; update layout and add error handling in GeneratingCaptionScreen

+++ app/src/main/java/com/example/capai_xml/data/dto/File.kt
@@ -1,7 +1,7 @@
 package com.example.capai_xml.data.dto
 
 data class File(
-    val audio_duration: Int,
+    val audio_duration: Double,
     val filename: String,
     val id: String,
     val number_of_channels: Int,
