+++ app/src/main/java/com/example/capai_xml/data/dto/Metadata.kt
@@ -1,8 +1,8 @@
 package com.example.capai_xml.data.dto
 
 data class Metadata(
-    val audio_duration: Int,
-    val billing_time: Int,
+    val audio_duration: Double,
+    val billing_time: Double,
     val number_of_distinct_channels: Int,
-    val transcription_time: Int
+    val transcription_time: Double
 )
