+++ app/src/main/java/com/example/capai_xml/data/dto/Moderation.kt
@@ -2,7 +2,7 @@ package com.example.capai_xml.data.dto
 
 data class Moderation(
     val error: Error,
-    val exec_time: Int,
+    val exec_time: Double,
     val is_empty: Boolean,
     val results: String,
     val success: Boolean
