+++ app/src/main/java/com/example/capai_xml/data/dto/ResultXX.kt
@@ -2,11 +2,11 @@ package com.example.capai_xml.data.dto
 
 data class ResultXX(
     val channel: Int,
-    val confidence: Int,
-    val end: Int,
+    val confidence: Double,
+    val end: Double,
     val language: String,
     val speaker: Int,
-    val start: Int,
+    val start: Double,
     val text: String,
     val words: List<Word>
 )
