+++ app/src/main/java/com/example/capai_xml/data/dto/Sentence.kt
@@ -2,7 +2,7 @@ package com.example.capai_xml.data.dto
 
 data class Sentence(
     val error: Error,
-    val exec_time: Int,
+    val exec_time: Double,
     val is_empty: Boolean,
     val results: List<String>,
     val success: Boolean
